requirements.txt
youtube-transcript-api>=0.6.0
spacy>=3.7.0
networkx>=3.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.0.0
pyyaml>=6.0
scipy>=1.11.0
